// types/supabase.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
// ...
  public: {
    Tables: {
      grade: {
        Row: {
          created_at: string;
          extra: Json | null;
          grade: number;
          id: string;
          subject: Database['public']['Enums']['grade_subject'];
          user_id: string;
        };
        Insert: {
          created_at?: string;
          extra?: Json | null;
          grade: number;
          id?: string;
          subject: Database['public']['Enums']['grade_subject'];
          user_id: string;
        };
        Update: {
          created_at?: string;
          extra?: Json | null;
          grade?: number;
          id?: string;
          subject?: Database['public']['Enums']['grade_subject'];
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'grade_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          }
        ];
      };
      user_data: {
        Row: {
          created_at: string;
          extra: Json | null;
          id: string;
          name: string;
          role: Database['public']['Enums']['user_data_role'];
          school_id: number;
        };
        Insert: {
          created_at?: string;
          extra?: Json | null;
          id: string;
          name: string;
          role: Database['public']['Enums']['user_data_role'];
          school_id: number;
        };
        Update: {
          created_at?: string;
          extra?: Json | null;
          id?: string;
          name?: string;
          role?: Database['public']['Enums']['user_data_role'];
          school_id?: number;
        };
        Relationships: [
          {
            foreignKeyName: 'user_data_id_fkey';
            columns: ['id'];
            isOneToOne: true;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          }
        ];
      };
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      grade_subject: 'MATH' | 'BIOLOGY' | 'ENGLISH' | 'PHYSICS' | 'CHEMISTRY' | 'HISTORY' | 'ECONOMICS';
      user_data_role: 'STUDENT' | 'TEACHER' | 'ADMIN';
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
// ...
};

type PublicSchema = Database[Extract<keyof Database, 'public'>];

export type Tables<
  PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] & PublicSchema['Views']) | { schema: keyof Database },
  TableName extends PublicTableNameOrOptions extends { schema: keyof Database }
    ? keyof (Database[PublicTableNameOrOptions['schema']]['Tables'] &
        Database[PublicTableNameOrOptions['schema']]['Views'])
    : never = never
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions['schema']]['Tables'] &
      Database[PublicTableNameOrOptions['schema']]['Views'])[TableName] extends {
      Row: infer R;
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema['Tables'] & PublicSchema['Views'])
    ? (PublicSchema['Tables'] & PublicSchema['Views'])[PublicTableNameOrOptions] extends {
        Row: infer R;
      }
      ? R
      : never
    : never;
// ...
export type Enums<
  PublicEnumNameOrOptions extends keyof PublicSchema['Enums'] | { schema: keyof Database },
  EnumName extends PublicEnumNameOrOptions extends { schema: keyof Database }
    ? keyof Database[PublicEnumNameOrOptions['schema']]['Enums']
    : never = never
> = PublicEnumNameOrOptions extends { schema: keyof Database }
  ? Database[PublicEnumNameOrOptions['schema']]['Enums'][EnumName]
  : PublicEnumNameOrOptions extends keyof PublicSchema['Enums']
    ? PublicSchema['Enums'][PublicEnumNameOrOptions]
    : never;